+++ zerver/lib/actions.py
@@ -7274,7 +7274,10 @@ def do_send_delete_user_group_event(realm: Realm, user_group_id: int, realm_id: 
 
 def check_delete_user_group(user_group_id: int, user_profile: UserProfile) -> None:
     user_group = access_user_group_by_id(user_group_id, user_profile)
-    user_group.delete()
+    if not user_group.active:
+        return
+    user_group.active = False
+    user_group.save(update_fields=["active"])
     do_send_delete_user_group_event(user_profile.realm, user_group_id, user_profile.realm.id)