+++ zerver/lib/user_groups.py
@@ -38,6 +38,7 @@ def user_groups_in_realm_serialized(realm: Realm) -> List[Dict[str, Any]]:
             name=user_group.name,
             description=user_group.description,
             members=[],
+            active=user_group.active,
         )
 
     membership = UserGroupMembership.objects.filter(user_group__realm=realm).values_list(